+++ src/gameobjects/particles/cParticle.h
@@ -12,9 +12,10 @@
 
 #pragma once
 
-#include "structs.h"
+// #include "structs.h"
 #include "utils/cRectangle.h"
 #include "utils/common.h"
+#include "gameobjects/particles/cParticleInfos.h"
 
 #include <map>
 #include <tuple>
